fix: fixing up loading tifs and 16 bit conversion and creating windows installer.
Fixed loading tifs to convert to 8 bit using np normalization.

Continued to convert to PIL RGB for compat with opencv. For some images,
they don't get converted to RGB during load. The number of channels is
checked in `preprocess_gray()` to ensure that opencv has the correct
images to work with

Created WindowsInstallerImproved.ps1 that (1) downloads the github zip
file with python files (2) download pixi (3) create a
`Run_Annotator.bat` that runs the pixi command in the project directory
and puts that on the desktop so the user gets a clickable icon to run
the software.